Try to use SendCommandAndIgnoreResponse when dispatching events
We are proposing to use SendCommandAndIgnoreResponse to dispatch event
to avoid waiting for the reply of sending events, so that we can
dispatch more than one events in one tick.

In order to control the time to send events we need to use tick
duration for every tick, it can come from pause action or pointermove
action. Adding it to pointermove is the simplest way to do it. We can
also add pause for every action.

Bug: 606367
Change-Id: Ib4745d37181de0724fe5e84fa32d893840ace910