Code health refactoring of animation.cc.
This patch aligns the implementation of web-animations with the spec.
Previously updates were performed via a PlayStateUpdateScope. This
class depended on an extra state (kPending) that is not to spec.
Support for microtasks is also introduced in this patch, which
addresses most of the remaining timing issues with the finished state
of an animation.

Bug: 960944
Change-Id: I62aff2254d2b5265f869415e0f763e140b460c37