[WPT/referrer-policy] Fix races around history.back()
Because history.back() works asynchronously,
there are races between history.back() and subsequent
promise_test()s.
For exmample:

1. The first 4K-length-referrer subtest finishes and history.back() is called.
2. The second subtest starts, and replaceState() is called.
3. The change to `location` caused by hisotory.back() in Step 1
   comes into effect.
4. The second subtest finishes -- at this time, `location` is the
   original Document URL
   (not '.../AAAA....AAA' set by replaceState() in Step 2)
   and thus history.back() has no effects.

This CL fixes the race by waiting for `popstate` events that should
be fired at the end of async tasks of history.back().

This CL reverts the changes to `referrer-policy-test-case.sub.js` in
https://github.com/web-platform-tests/wpt/pull/18401
https://github.com/web-platform-tests/wpt/commit/8b46cbf469fd282e56ea3ef883e8a8cb509d49d6
and waits for popstate instead, because single asyncResolve() isn't sufficient
to fix the race condition described above.

Bug: 906850
Change-Id: I21e7b8dd6001a13230fd8dca4d1693ae956958ed