Create Web Platform Tests for Scroll to text
Add a web platform test that performs navigations to a test page with
various targetText parameters and checks if the target page
successfully scrolled as expected. Using a BroadcastChannel is the only
way to communicate whether the target page scrolled, since scroll to
text is specifically restricted from iframes or pages with an opener,
so there's no other way for the test page to track the status of the
child target page.

Bug: 994299
Change-Id: I69243e739c3a7469ac48647508e379f204ccfbf6